added first part of debug errors and changed order to push pending data

+++ EQ/EQNewOrderViewModel.h
@@ -32,19 +32,19 @@
 @property (nonatomic,strong) NSArray *group2;
 @property (nonatomic,strong) NSArray *categories;
 
-@property (nonatomic,assign) int categorySelected;
-@property (nonatomic,assign) int group1Selected;
-@property (nonatomic,assign) int group2Selected;
-@property (nonatomic,assign) int articleSelectedIndex;
+@property (nonatomic,assign) NSUInteger categorySelected;
+@property (nonatomic,assign) NSUInteger group1Selected;
+@property (nonatomic,assign) NSUInteger group2Selected;
+@property (nonatomic,assign) NSUInteger articleSelectedIndex;
 
 @property (nonatomic,assign) BOOL newOrder;
 
-- (void)defineSelectedCategory:(int)index;
-- (void)defineSelectedGroup1:(int)index;
-- (void)defineSelectedGroup2:(int)index;
-- (void)defineSelectedArticle:(int)index;
-- (void)defineOrderStatus:(int)index;
-- (BOOL)addItemQuantity:(int)quantity;
+- (void)defineSelectedCategory:(NSUInteger)index;
+- (void)defineSelectedGroup1:(NSUInteger)index;
+- (void)defineSelectedGroup2:(NSUInteger)index;
+- (void)defineSelectedArticle:(NSUInteger)index;
+- (void)defineOrderStatus:(NSUInteger)index;
+- (BOOL)addItemQuantity:(NSUInteger)quantity;
 - (NSNumber *)itemsQuantity;
 - (NSNumber *)subTotal;
 - (float)discountPercentage;
@@ -57,9 +57,9 @@
 - (void)editItem:(ItemPedido *)item;
 - (NSNumber *)quantityOfCurrentArticle;
 - (void)cancelOrder;
-- (void)sortArticlesByIndex:(int)index;
-- (void)sortGroup2ByIndex:(int)index;
-- (void)sortGroup1ByIndex:(int)index;
+- (void)sortArticlesByIndex:(NSUInteger)index;
+- (void)sortGroup2ByIndex:(NSUInteger)index;
+- (void)sortGroup1ByIndex:(NSUInteger)index;
 
 - (NSString *)orderHTML;
